Update on "[refactor] Move ext_arr/any_arr/template to taichi.type.annotations."
[ghstack-poisoned]
