+++ python/taichi/ui/canvas.py
@@ -1,8 +1,8 @@
 from taichi.core import ti_core as _ti_core
 from taichi.lang.impl import default_cfg
-from taichi.lang.kernel_arguments import ext_arr, template
 from taichi.lang.kernel_impl import kernel
 from taichi.lang.ops import get_addr
+from taichi.type.annotations import ext_arr, template
 
 from .staging_buffer import (copy_colors_to_vbo, copy_vertices_to_vbo,
                              get_vbo_field, to_u8_rgba)
