+++ python/taichi/ui/gui.py
@@ -3,9 +3,9 @@ from contextlib import contextmanager
 
 from taichi.core import ti_core as _ti_core
 from taichi.lang.impl import default_cfg
-from taichi.lang.kernel_arguments import ext_arr, template
 from taichi.lang.kernel_impl import kernel
 from taichi.lang.ops import get_addr
+from taichi.type.annotations import ext_arr, template
 
 from .utils import *
